Trim the end of the body
For some messages that are essentially duplicates, they may have
some varying extra line endings at the end of the body.
Otherwise, they are the same, so best to trim for comparison purposes.
Note, this only trims a copy the body of the MailItem being checked,
not the MailItem body itself.